+++ ambry-cloud/src/main/java/com.github.ambry.cloud/CloudDestination.java
@@ -40,7 +40,7 @@ public interface CloudDestination {
   /**
    * Download blob from the cloud destination.
    * @param blobId id of the Ambry blob to be downloaded
-   * @return BlobReadInfo that has the data and metadata for the blob being downloaded
+   * @param outputStream outputstream to populate the downloaded data with
    * @throws CloudStorageException if the download encounters an error.
    */
   void downloadBlob(BlobId blobId, OutputStream outputStream) throws CloudStorageException;
